move no_packet_loops to avoid tunnel conflict
